+++ Modules/_testinternalcapi.c
@@ -1210,7 +1210,7 @@ write_perf_map_entry(PyObject *self, PyObject *args)
 {
     PyObject *code_addr_v;
     const void *code_addr;
-    unsigned int code_size;
+    size_t code_size;
     const char *entry_name;
 
     if (!PyArg_ParseTuple(args, "OIs", &code_addr_v, &code_size, &entry_name))
